Move TestLanguageServiceHost to OverlayLanguageServiceHost
